core: fix some clang-tidy warnings

+++ core/src/screenshot.cpp
@@ -2,7 +2,7 @@
 
 Screenshot::Screenshot(int width) {
   assert(width > 0 && "Can't create a Screenshot with zero width.");
-  drawingOptions.windowWidth = width;
+  drawingOptions.windowWidth = float(width);
 }
 
 Screenshot::~Screenshot() {
@@ -16,7 +16,7 @@ std::string Screenshot::getBase64Screenshot(Scene *scene) {
   if (!canvas) {
     constexpr auto viewHeightToWidthRatio = 7.0f / 5.0f;
     canvas = love::DrawDataFrame::newCanvas(
-        drawingOptions.windowWidth, drawingOptions.windowWidth * viewHeightToWidthRatio);
+        int(drawingOptions.windowWidth), int(drawingOptions.windowWidth * viewHeightToWidthRatio));
   }
   love::DrawDataFrame::renderToCanvas(canvas, [scene, this]() {
     scene->draw(drawingOptions);
